requirements.txt
alembic==1.13.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bayesian-optimization==1.4.3
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
dask==2023.5.0
DateTools==1.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.15.4
fonttools==4.50.0
forex-python==1.8
fqdn==1.5.1
frozendict==2.4.4
fsspec==2024.6.1
greenlet==3.0.3
html5lib==1.1
idna==3.6
imbalanced-learn==0.12.3
imblearn==0.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipykernel==5.5.5
ipyparallel==8.8.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-black==0.3.4
jupyter-console==6.4.4
jupyter-events==0.9.0
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab_code_formatter==2.2.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
locket==1.0.0
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
mizani==0.9.3
multitasking==0.0.11
mwparserfromhell==0.6.6
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==6.5.6
notebook_shim==0.2.4
numpy==1.24.4
openpyxl==3.1.2
optuna==4.0.0
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
partd==1.4.1
pathspec==0.12.1
patsy==0.5.6
peewee==3.17.6
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.20.0
plotnine==0.12.4
powershell_kernel==0.1.4
progressbar2==4.4.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.21
pygam==0.9.1
Pygments==2.17.2
PyMySQL==1.1.0
pyodbc==5.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-settings==0.2.2
python-utils==3.8.2
pytz==2024.1
pywikibot==9.0.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
statsmodels==0.14.1
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tldextract==5.1.2
tokenize-rt==5.2.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
xgboost==2.1.1
yfinance==0.2.41
zipp==3.17.0
